MergeSort

Documentation

halve

Arguments

:: [x]
The input list

-> ([x], [x])
The tuple containing two lists

Split a list into two halves. If the length is odd, the first returned half will be shorter (rounds down)

merge

Arguments

:: Ord x

=> ([x], [x])
The input tuple containing two lists

-> [x]
The merged list

Merges split lists into single sorted list. Compares each element of both lists and inserts values smallest to largest

mergeSort

Arguments

:: Ord x

=> [x]
The input list to be sorted

-> [x]
The sorted list

Performs Merge Sort on single list